Rename: hooks 폴더 정리
