+++ src/components/common/detail/Comment/NewComment.tsx
@@ -5,7 +5,7 @@ import Swal from 'sweetalert2';
 import { NewIcon } from '@components/atoms/Icon/NewIcon';
 import { Line } from '@components/atoms/Line/Line';
 import { CommentSubmitArea } from 'pages/DetailPage';
-import { useDeleteComment, useDeleteReply } from '@hooks/useMutateQuery';
+import { useDeleteComment, useDeleteReply } from '@hooks/query/useMutateQuery';
 import { UserState } from '@store/index';
 import { createMarkup } from '@utils/createMarkup';
 import { handleErrorType } from '@utils/handleErrorType';
