+++ src/components/common/detail/CommentInput/CommentInput.tsx
@@ -4,7 +4,7 @@ import React, { useState } from 'react';
 import { useRecoilState } from 'recoil';
 import classes from './commentInput.module.scss';
 import { Button, Input } from '@components/atoms';
-import { useAddComment, useAddReply } from '@hooks/useMutateQuery';
+import { useAddComment, useAddReply } from '@hooks/query/useMutateQuery';
 interface ICommentInputPropsType {
   postId?: number;
   commentId?: number;
