+++ src/components/common/join/Modal/JoinModal.tsx
@@ -2,7 +2,10 @@ import React, { ChangeEvent, Dispatch, SetStateAction, useState } from 'react';
 import { Button, Modal } from '@components/atoms';
 import './joinModal.scss';
 // import { useParams } from 'react-router-dom';
-import { useDeleteApplicant, useUpdateApplicant } from '@hooks/useMutateQuery';
+import {
+  useDeleteApplicant,
+  useUpdateApplicant,
+} from '@hooks/query/useMutateQuery';
 import { useRecoilValue } from 'recoil';
 import { UserState } from '@store/user';
