+++ src/components/common/join/OwnerJoinerCheck/OwnerJoinerCheck.tsx
@@ -1,7 +1,8 @@
-import { IApplicantDetail } from '@api/responseType';
+// import { IApplicantDetail } from '@api/responseType';
+import { IApplicantDetail } from 'types/api.types';
 import { SkillIcon } from '@components/atoms/SkillIcon/SkillIcon';
-import { useGetApplicant, useGetPostByPostId } from '@hooks/useGetQuery';
-import React, { useState } from 'react';
+import { useGetApplicant, useGetPostByPostId } from '@hooks/query/useGetQuery';
+import { useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { ViewModal } from '../Modal/ViewModal';
 import './ownerJoinerCheck.scss';
