+++ src/components/common/main/AlarmIcon/AlarmItem.tsx
@@ -3,7 +3,7 @@ import classes from './alarmIcon.module.scss';
 import { useNavigate } from 'react-router-dom';
 import { AlarmType } from 'types/api.types';
 import { handleTime } from '@utils/handleTime';
-import { useUpdateAlarm } from '@hooks/useMutateQuery';
+import { useUpdateAlarm } from '@hooks/query/useMutateQuery';
 
 interface Props {
   alarm: AlarmType;
