+++ src/components/common/onboard/OnboardModal/OnboardModal.tsx
@@ -4,7 +4,7 @@ import { Alert, Button, Modal } from '@components/atoms';
 import classes from './onboardmodal.module.scss';
 import AddProfileModal from './AddProfileModal';
 import CheckPositionModal from './CheckPositionModal';
-import { useAddUserInfo } from '@hooks/useMutateQuery';
+import { useAddUserInfo } from '@hooks/query/useMutateQuery';
 import useProfileImage from '@hooks/useProfileImage';
 
 interface IOnboardModalProps {
