+++ src/hooks/query/useInfiniteQueryCarousel.tsx
